pyproject.toml
[project]
name = "VideoArchive"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "torch",
    "transformers",
    "Pillow",
    "opencv-python-headless",
    "faiss-cpu",
    "numpy",
    "tqdm",
    "scenedetect>=0.6.7",
]
